build : generate empty folders for frame the folder structure
