+++ src/constants/paymentState.js
@@ -1,0 +1,6 @@
+export const ORDER_STATE = {
+  READY: "READY",
+  SUCCESS: "SUCCESS",
+  PAID: "PAID",
+  CANCEL: "CANCEL",
+};
